# api-docs.json
{"openapi":"3.0.1","info":{"title":"T1 Open School AOP","description":"Sample API of the AOP Test","contact":{"name":"Karabanov Andrey","email":"deskpa@yandex.ru"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Default Server URL"}],"tags":[{"name":"Methods Data API","description":"API for methods data"},{"name":"Film API","description":"API for films"}],"paths":{"/film":{"get":{"tags":["Film API"],"summary":"Get all films","description":"Get a list of all films.","operationId":"getAllFilms","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"post":{"tags":["Film API"],"summary":"Create a new film","description":"Creates a new film","operationId":"createFilm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"/methodData":{"get":{"tags":["Methods Data API"],"summary":"Get all methods data","description":"Get a list of all  methods data.","operationId":"getAllMethodsData","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MethodDataDTO"}}}}}}},"delete":{"tags":["Methods Data API"],"summary":"Delete all methods data","description":"Delete all methods data ","operationId":"deleteAllMethodsData","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/methodData/stat/{methodName}":{"get":{"tags":["Methods Data API"],"summary":"Get statistic from method data by method name","description":"Get statistic from all methods data by its name.","operationId":"getStatByMethodName","parameters":[{"name":"methodName","in":"path","description":"Name of the method to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MethodDataStatDTO"}}}}}}},"/methodData/methodName/{methodName}":{"get":{"tags":["Methods Data API"],"summary":"Get all method data by method name","description":"Get all methods data by its name.","operationId":"getAllMethodsDataByMethodName","parameters":[{"name":"methodName","in":"path","description":"Name of the method to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MethodDataDTO"}}}}}}}},"/film/titleName/{titleName}":{"get":{"tags":["Film API"],"summary":"Get film by title","description":"Get a film by its title name.","operationId":"getFilmByTitleName","parameters":[{"name":"titleName","in":"path","description":"Title name of the film to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"/film/producer/{producer}":{"get":{"tags":["Film API"],"summary":"Get film by producer","description":"Get a film by its producer.","operationId":"getFilmByProducer","parameters":[{"name":"producer","in":"path","description":"Producer of the film to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"/film/id/{id}":{"get":{"tags":["Film API"],"summary":"Get film by ID","description":"Get a film by its ID.","operationId":"getFilmById","parameters":[{"name":"id","in":"path","description":"ID of the film to be obtained","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"/film/genre/{genre}":{"get":{"tags":["Film API"],"summary":"Get film by genre","description":"Get a film by its genre.","operationId":"getFilmByGenre","parameters":[{"name":"genre","in":"path","description":"Genre of the film to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}},"/film/country/{country}":{"get":{"tags":["Film API"],"summary":"Get film by country","description":"Get a film by its country.","operationId":"getFilmByCountry","parameters":[{"name":"country","in":"path","description":"Country of the film to be obtained","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmDTO"}}}}}}}},"components":{"schemas":{"FilmDTO":{"type":"object","properties":{"id":{"type":"integer","description":"Unique identifier of the film","format":"int64","readOnly":true,"example":1},"titleName":{"type":"string","description":"Title name of the film","example":"Catch Me If You Can"},"genre":{"type":"string","description":"Genre of the film","example":"Comedy"},"country":{"type":"string","description":"Country of the film","example":"United States"},"producer":{"type":"string","description":"Producer of the film","example":"Steven Spielberg"}},"description":"DTO for Film entity"},"MethodDataDTO":{"type":"object","properties":{"executeDate":{"type":"string","description":"Execution date of the method","format":"date-time","example":"2024-05-04T23:29:18.525Z"},"methodName":{"type":"string","description":"Name of the method","example":"FilmServiceImpl.create(..)"},"executeNanoTime":{"type":"integer","description":"Execution time of the method (nano)","format":"int64","example":49},"executeMilliTime":{"type":"integer","description":"Execution time of the method (ms)","format":"int64","example":49},"annotationType":{"type":"string","description":"Annotation type of the method","example":"ASYNC","enum":["DEFAULT","ASYNC"]}},"description":"DTO for methods data entity"},"MethodDataStatDTO":{"type":"object","properties":{"methodName":{"type":"string","description":"Name of the method"},"minExecuteTimeMilli":{"type":"integer","description":"Minimum method execution time (ms)","format":"int64","example":1},"maxExecuteTimeMilli":{"type":"integer","description":"Maximum method execution time (ms)","format":"int64","example":42},"avgExecuteTimeMilli":{"type":"integer","description":"Execution time of the method (ms)","format":"int64","example":49},"minExecuteTimeNano":{"type":"integer","description":"Minimum method execution time (nano)","format":"int64","example":112312},"maxExecuteTimeNano":{"type":"integer","description":"Maximum method execution time (nano)","format":"int64","example":1903121},"avgExecuteTimeNano":{"type":"integer","description":"Execution time of the method (nano)","format":"int64","example":18156778},"annotationType":{"type":"string","description":"Annotation type of the method","example":"ASYNC","enum":["DEFAULT","ASYNC"]}},"description":"DTO for method data statistics"}}}}